Fix IAR build issues (#2533)
* Fix IAR build issues

* Address review comments

+++ libraries/c_sdk/standard/mqtt/test/unit/iot_tests_mqtt_api.c
@@ -544,9 +544,9 @@ static int32_t transportSend( const NetworkContext_t * pNetworkContext,
 /**
  * @brief A transport send function that delays.
  */
-static size_t transportSendDelay( const NetworkContext_t * pSendContext,
-                                  const void * pMessage,
-                                  size_t messageLength )
+static int32_t transportSendDelay( const NetworkContext_t * pSendContext,
+                                   const void * pMessage,
+                                   size_t messageLength )
 {
     IotSemaphore_t * pWaitSem = ( IotSemaphore_t * ) pSendContext;
 
@@ -560,7 +560,7 @@ static size_t transportSendDelay( const NetworkContext_t * pSendContext,
     IotClock_SleepMs( 2000 );
 
     /* This function returns the message length to simulate a successful send. */
-    return messageLength;
+    return ( int32_t ) messageLength;
 }
 
 /*-----------------------------------------------------------*/
